requirements.txt
azure-core==1.29.4
azure-storage-blob==12.8.1
azure-identity==1.14.1
pymp4==1.4.0
pycountry==22.3.5